+++ src/Fable.Transforms/Replacements.fs
@@ -1565,7 +1565,12 @@ let strings (com: ICompiler) (ctx: Context) r t (i: CallInfo) (thisArg: Expr opt
 
             Helper.LibCall(com, "String", methName, t, c :: args, hasSpread = spread, ?loc = r)
             |> Some
-    | "ToCharArray", Some c, _ -> stringToCharArray c |> Some
+    | "ToCharArray", Some c, _ ->
+        match args with
+        | [] -> stringToCharArray c |> Some
+        | [ ExprType(Number(Int32, _)); ExprType(Number(Int32, _)) ] ->
+            Helper.LibCall(com, "String", "toCharArray2", t, c :: args, ?loc = r) |> Some
+        | _ -> None
     | "Split", Some c, _ ->
         match args with
         // Optimization